Add (and use) new client up endpoint
The general goal of this is that a client can just hit a single
well-known endpoint when coming up, and use that to discover what URLs
to hit for other actions, instead of having to know many.

This offers overkill levels of flexibility on the backend.

Changes here entail:

- Add API endpoint
- Remove some unused imports
- Add a websockets module (unrelated, somewhat random move)
- Move some things around in patchbay module (_slightly_ better encapsulation)

Should probably also:

- Deprecate (or just remove) the old endpoint
- Fix some sorta-whacky variable names in the Arduino sketch